+++ src/app/test/page.tsx
@@ -1,0 +1,7 @@
+import { Header } from '@/src/components/common/header';
+
+export default function Test() {
+    return (
+        <Header />
+    );
+}